+++ changelog.md
@@ -1,1 +1,2 @@
-- Added french translation
+- Fixed empty ItemStack getting NBT data
+- Fixed can't take album out of empty album in lectern